[TM-702] Share some transition logic between the two entity-related state machines.
