+++ app/Models/Traits/HasEntityStatus.php
@@ -5,7 +5,6 @@ namespace App\Models\Traits;
 use App\Events\V2\General\EntityStatusChangeEvent;
 use App\StateMachines\EntityStatusStateMachine;
 use Asantibanez\LaravelEloquentStateMachines\Traits\HasStateMachines;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Str;
 
 /**
@@ -21,6 +20,7 @@ use Illuminate\Support\Str;
 trait HasEntityStatus {
     use HasStatus;
     use HasStateMachines;
+    use HasEntityStatusScopesAndTransitions;
 
     public $stateMachines = [
         'status' => EntityStatusStateMachine::class,
@@ -32,42 +32,6 @@ trait HasEntityStatus {
         EntityStatusStateMachine::NEEDS_MORE_INFORMATION => 'Needs more information',
         EntityStatusStateMachine::APPROVED => 'Approved',
     ];
-
-    public function scopeIsApproved(Builder $query): Builder
-    {
-        return $this->scopeIsStatus($query, EntityStatusStateMachine::APPROVED);
-    }
-
-    public function isEditable(): bool
-    {
-        return $this->status == EntityStatusStateMachine::STARTED;
-    }
-
-    public function approve($feedback): void
-    {
-        $this->feedback = $feedback;
-        $this->feedback_fields = null;
-
-        if ($this->status == EntityStatusStateMachine::APPROVED) {
-            // If we were already approved, this may have been called because an update request got approved, and
-            // we need to make sure the transition hooks execute, so fake us into awaiting-approval first.
-            $this->status = EntityStatusStateMachine::AWAITING_APPROVAL;
-        }
-
-        $this->status()->transitionTo(EntityStatusStateMachine::APPROVED);
-    }
-
-    public function submitForApproval(): void
-    {
-        $this->status()->transitionTo(EntityStatusStateMachine::AWAITING_APPROVAL);
-    }
-
-    public function needsMoreInformation($feedback, $feedbackFields): void
-    {
-        $this->feedback = $feedback;
-        $this->feedback_fields = $feedbackFields;
-        $this->status()->transitionTo(EntityStatusStateMachine::NEEDS_MORE_INFORMATION);
-    }
 
     public function dispatchStatusChangeEvent($user): void
     {
